move BillingEnv-creating code from env.js to stripe.js and billing.js
